different vars types for 2022 sample

+++ include/triggers.hxx
@@ -10,9 +10,9 @@ bool matchParticle(const ROOT::Math::PtEtaPhiMVector &particle,
                    ROOT::RVec<float> &triggerobject_etas,
                    ROOT::RVec<float> &triggerobject_phis,
                    ROOT::RVec<int> &triggerobject_bits,
-                   ROOT::RVec<int> &triggerobject_ids, const float &matchDeltaR,
+                   ROOT::RVec<unsigned short> &triggerobject_ids, const float &matchDeltaR,
                    const float &pt_cut, const float &eta_cut,
-                   const int &trigger_particle_id_cut,
+                   const unsigned short &trigger_particle_id_cut,
                    const int &triggerbit_cut);
 
 ROOT::RDF::RNode GenerateSingleTriggerFlag(
@@ -21,7 +21,7 @@ ROOT::RDF::RNode GenerateSingleTriggerFlag(
     const std::string &triggerobject_id, const std::string &triggerobject_pt,
     const std::string &triggerobject_eta, const std::string &triggerobject_phi,
     const std::string &hltpath, const float &pt_cut, const float &eta_cut,
-    const int &trigger_particle_id_cut, const int &triggerbit_cut,
+    const unsigned short &trigger_particle_id_cut, const int &triggerbit_cut,
     const float &DeltaR_threshold);
 
 ROOT::RDF::RNode GenerateDoubleTriggerFlag(
@@ -31,8 +31,8 @@ ROOT::RDF::RNode GenerateDoubleTriggerFlag(
     const std::string &triggerobject_pt, const std::string &triggerobject_eta,
     const std::string &triggerobject_phi, const std::string &hltpath,
     const float &p1_pt_cut, const float &p2_pt_cut, const float &p1_eta_cut,
-    const float &p2_eta_cut, const int &p1_trigger_particle_id_cut,
-    const int &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
+    const float &p2_eta_cut, const unsigned short &p1_trigger_particle_id_cut,
+    const unsigned short &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
     const int &p2_triggerbit_cut, const float &DeltaR_threshold);
 
 ROOT::RDF::RNode GenerateTripleTriggerORFlag(
@@ -42,8 +42,8 @@ ROOT::RDF::RNode GenerateTripleTriggerORFlag(
     const std::string &triggerobject_pt, const std::string &triggerobject_eta,
     const std::string &triggerobject_phi, const std::string &hltpath,
     const float &p1_pt_cut, const float &p2_pt_cut, const float &p3_pt_cut, const float &p1_eta_cut,
-    const float &p2_eta_cut, const float &p3_eta_cut, const int &p1_trigger_particle_id_cut,
-    const int &p2_trigger_particle_id_cut, const int &p3_trigger_particle_id_cut, const int &p1_triggerbit_cut,
+    const float &p2_eta_cut, const float &p3_eta_cut, const unsigned short &p1_trigger_particle_id_cut,
+    const unsigned short &p2_trigger_particle_id_cut, const unsigned short &p3_trigger_particle_id_cut, const int &p1_triggerbit_cut,
     const int &p2_triggerbit_cut, const int &p3_triggerbit_cut, const float &DeltaR_threshold);
 
 ROOT::RDF::RNode GenerateDoubleTriggerORFlag(
@@ -53,8 +53,8 @@ ROOT::RDF::RNode GenerateDoubleTriggerORFlag(
     const std::string &triggerobject_pt, const std::string &triggerobject_eta,
     const std::string &triggerobject_phi, const std::string &hltpath,
     const float &p1_pt_cut, const float &p2_pt_cut, const float &p1_eta_cut,
-    const float &p2_eta_cut,  const int &p1_trigger_particle_id_cut,
-    const int &p2_trigger_particle_id_cut,  const int &p1_triggerbit_cut,
+    const float &p2_eta_cut,  const unsigned short &p1_trigger_particle_id_cut,
+    const unsigned short &p2_trigger_particle_id_cut,  const int &p1_triggerbit_cut,
     const int &p2_triggerbit_cut, const float &DeltaR_threshold);
 
 ROOT::RDF::RNode GenerateQuadTriggerORFlag(
@@ -65,7 +65,7 @@ ROOT::RDF::RNode GenerateQuadTriggerORFlag(
     const std::string &triggerobject_phi, const std::string &hltpath,
     const float &p1_pt_cut, const float &p2_pt_cut, const float &p3_pt_cut, const float &p4_pt_cut,
     const float &p1_eta_cut, const float &p2_eta_cut, const float &p3_eta_cut, const float &p4_eta_cut,
-    const int &p1_trigger_particle_id_cut, const int &p2_trigger_particle_id_cut, const int &p3_trigger_particle_id_cut, const int &p4_trigger_particle_id_cut,
+    const unsigned short &p1_trigger_particle_id_cut, const unsigned short &p2_trigger_particle_id_cut, const unsigned short &p3_trigger_particle_id_cut, const unsigned short &p4_trigger_particle_id_cut,
     const int &p1_triggerbit_cut, const int &p2_triggerbit_cut, const int &p3_triggerbit_cut, const int &p4_triggerbit_cut, const float &DeltaR_threshold);
 
 ROOT::RDF::RNode MatchSingleTriggerObject(
@@ -74,7 +74,7 @@ ROOT::RDF::RNode MatchSingleTriggerObject(
     const std::string &triggerobject_id, const std::string &triggerobject_pt,
     const std::string &triggerobject_eta, const std::string &triggerobject_phi,
     const float &pt_cut, const float &eta_cut,
-    const int &trigger_particle_id_cut, const int &triggerbit_cut,
+    const unsigned short &trigger_particle_id_cut, const int &triggerbit_cut,
     const float &DeltaR_threshold);
 
 ROOT::RDF::RNode MatchDoubleTriggerObject(
@@ -84,8 +84,8 @@ ROOT::RDF::RNode MatchDoubleTriggerObject(
     const std::string &triggerobject_pt, const std::string &triggerobject_eta,
     const std::string &triggerobject_phi, const float &p1_pt_cut,
     const float &p2_pt_cut, const float &p1_eta_cut, const float &p2_eta_cut,
-    const int &p1_trigger_particle_id_cut,
-    const int &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
+    const unsigned short &p1_trigger_particle_id_cut,
+    const unsigned short &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
     const int &p2_triggerbit_cut, const float &DeltaR_threshold);
 
 ROOT::RDF::RNode GetPrescaleValues(ROOT::RDF::RNode df,
@@ -102,9 +102,9 @@ bool matchParticle(const ROOT::Math::PtEtaPhiMVector &particle,
                    ROOT::RVec<float> &triggerobject_etas,
                    ROOT::RVec<float> &triggerobject_phis,
                    ROOT::RVec<int> &triggerobject_bits,
-                   ROOT::RVec<int> &triggerobject_ids, const float &matchDeltaR,
+                   ROOT::RVec<unsigned short> &triggerobject_ids, const float &matchDeltaR,
                    const float &pt_cut, const float &eta_cut,
-                   const int &trigger_particle_id_cut,
+                   const unsigned short &trigger_particle_id_cut,
                    const int &triggerbit_cut,
                    const float &trigger_particle_pt_cut);
 
@@ -114,7 +114,7 @@ ROOT::RDF::RNode GenerateSingleTriggerFlag(
     const std::string &triggerobject_id, const std::string &triggerobject_pt,
     const std::string &triggerobject_eta, const std::string &triggerobject_phi,
     const std::string &hltpath, const float &pt_cut, const float &eta_cut,
-    const int &trigger_particle_id_cut, const int &triggerbit_cut,
+    const unsigned short &trigger_particle_id_cut, const int &triggerbit_cut,
     const float &DeltaR_threshold, const float &trigger_particle_pt_cut);
 
 ROOT::RDF::RNode GenerateDoubleTriggerFlag(
@@ -124,8 +124,8 @@ ROOT::RDF::RNode GenerateDoubleTriggerFlag(
     const std::string &triggerobject_pt, const std::string &triggerobject_eta,
     const std::string &triggerobject_phi, const std::string &hltpath,
     const float &p1_pt_cut, const float &p2_pt_cut, const float &p1_eta_cut,
-    const float &p2_eta_cut, const int &p1_trigger_particle_id_cut,
-    const int &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
+    const float &p2_eta_cut, const unsigned short &p1_trigger_particle_id_cut,
+    const unsigned short &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
     const int &p2_triggerbit_cut, const float &DeltaR_threshold,
     const float &p1_trigger_particle_pt_cut,
     const float &p2_trigger_particle_pt_cut);
