+++ src/triggers.cxx
@@ -116,9 +116,9 @@ bool matchParticle(const ROOT::Math::PtEtaPhiMVector &particle,
                    ROOT::RVec<float> &triggerobject_etas,
                    ROOT::RVec<float> &triggerobject_phis,
                    ROOT::RVec<int> &triggerobject_bits,
-                   ROOT::RVec<int> &triggerobject_ids, const float &matchDeltaR,
+                   ROOT::RVec<unsigned short> &triggerobject_ids, const float &matchDeltaR,
                    const float &pt_cut, const float &eta_cut,
-                   const int &trigger_particle_id_cut,
+                   const unsigned short &trigger_particle_id_cut,
                    const int &triggerbit_cut) {
     Logger::get("CheckTriggerMatch")->debug("Checking Triggerobjects");
     Logger::get("CheckTriggerMatch")
@@ -214,7 +214,7 @@ ROOT::RDF::RNode GenerateSingleTriggerFlag(
     const std::string &triggerobject_id, const std::string &triggerobject_pt,
     const std::string &triggerobject_eta, const std::string &triggerobject_phi,
     const std::string &hltpath, const float &pt_cut, const float &eta_cut,
-    const int &trigger_particle_id_cut, const int &triggerbit_cut,
+    const unsigned short &trigger_particle_id_cut, const int &triggerbit_cut,
     const float &DeltaR_threshold) {
 
     auto triggermatch =
@@ -222,7 +222,7 @@ ROOT::RDF::RNode GenerateSingleTriggerFlag(
          triggerbit_cut](bool hltpath,
                          const ROOT::Math::PtEtaPhiMVector &particle_p4,
                          ROOT::RVec<int> triggerobject_bits,
-                         ROOT::RVec<int> triggerobject_ids,
+                         ROOT::RVec<unsigned short> triggerobject_ids,
                          ROOT::RVec<float> triggerobject_pts,
                          ROOT::RVec<float> triggerobject_etas,
                          ROOT::RVec<float> triggerobject_phis) {
@@ -343,8 +343,8 @@ ROOT::RDF::RNode GenerateDoubleTriggerFlag(
     const std::string &triggerobject_pt, const std::string &triggerobject_eta,
     const std::string &triggerobject_phi, const std::string &hltpath,
     const float &p1_pt_cut, const float &p2_pt_cut, const float &p1_eta_cut,
-    const float &p2_eta_cut, const int &p1_trigger_particle_id_cut,
-    const int &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
+    const float &p2_eta_cut, const unsigned short &p1_trigger_particle_id_cut,
+    const unsigned short &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
     const int &p2_triggerbit_cut, const float &DeltaR_threshold) {
 
     auto triggermatch = [DeltaR_threshold, p1_pt_cut, p2_pt_cut, p1_eta_cut,
@@ -355,7 +355,7 @@ ROOT::RDF::RNode GenerateDoubleTriggerFlag(
                             const ROOT::Math::PtEtaPhiMVector &particle1_p4,
                             const ROOT::Math::PtEtaPhiMVector &particle2_p4,
                             ROOT::RVec<int> triggerobject_bits,
-                            ROOT::RVec<int> triggerobject_ids,
+                            ROOT::RVec<unsigned short> triggerobject_ids,
                             ROOT::RVec<float> triggerobject_pts,
                             ROOT::RVec<float> triggerobject_etas,
                             ROOT::RVec<float> triggerobject_phis) {
@@ -499,8 +499,8 @@ ROOT::RDF::RNode GenerateTripleTriggerORFlag(
     const std::string &triggerobject_pt, const std::string &triggerobject_eta,
     const std::string &triggerobject_phi, const std::string &hltpath,
     const float &p1_pt_cut, const float &p2_pt_cut, const float &p3_pt_cut, const float &p1_eta_cut,
-    const float &p2_eta_cut, const float &p3_eta_cut, const int &p1_trigger_particle_id_cut,
-    const int &p2_trigger_particle_id_cut, const int &p3_trigger_particle_id_cut, const int &p1_triggerbit_cut,
+    const float &p2_eta_cut, const float &p3_eta_cut, const unsigned short &p1_trigger_particle_id_cut,
+    const unsigned short &p2_trigger_particle_id_cut, const unsigned short &p3_trigger_particle_id_cut, const int &p1_triggerbit_cut,
     const int &p2_triggerbit_cut, const int &p3_triggerbit_cut, const float &DeltaR_threshold) {
 
     auto triggermatch = [DeltaR_threshold, p1_pt_cut, p2_pt_cut, p3_pt_cut, p1_eta_cut,
@@ -512,7 +512,7 @@ ROOT::RDF::RNode GenerateTripleTriggerORFlag(
                             const ROOT::Math::PtEtaPhiMVector &particle2_p4,
                             const ROOT::Math::PtEtaPhiMVector &particle3_p4,
                             ROOT::RVec<int> triggerobject_bits,
-                            ROOT::RVec<int> triggerobject_ids,
+                            ROOT::RVec<unsigned short> triggerobject_ids,
                             ROOT::RVec<float> triggerobject_pts,
                             ROOT::RVec<float> triggerobject_etas,
                             ROOT::RVec<float> triggerobject_phis) {
@@ -600,8 +600,8 @@ ROOT::RDF::RNode GenerateDoubleTriggerORFlag(
     const std::string &triggerobject_pt, const std::string &triggerobject_eta,
     const std::string &triggerobject_phi, const std::string &hltpath,
     const float &p1_pt_cut, const float &p2_pt_cut, const float &p1_eta_cut,
-    const float &p2_eta_cut, const int &p1_trigger_particle_id_cut,
-    const int &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
+    const float &p2_eta_cut, const unsigned short &p1_trigger_particle_id_cut,
+    const unsigned short &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
     const int &p2_triggerbit_cut, const float &DeltaR_threshold) {
 
     auto triggermatch = [DeltaR_threshold, p1_pt_cut, p2_pt_cut, p1_eta_cut,
@@ -612,7 +612,7 @@ ROOT::RDF::RNode GenerateDoubleTriggerORFlag(
                             const ROOT::Math::PtEtaPhiMVector &particle1_p4,
                             const ROOT::Math::PtEtaPhiMVector &particle2_p4,
                             ROOT::RVec<int> triggerobject_bits,
-                            ROOT::RVec<int> triggerobject_ids,
+                            ROOT::RVec<unsigned short> triggerobject_ids,
                             ROOT::RVec<float> triggerobject_pts,
                             ROOT::RVec<float> triggerobject_etas,
                             ROOT::RVec<float> triggerobject_phis) {
@@ -692,7 +692,7 @@ ROOT::RDF::RNode GenerateQuadTriggerORFlag(
     const std::string &triggerobject_phi, const std::string &hltpath,
     const float &p1_pt_cut, const float &p2_pt_cut, const float &p3_pt_cut, const float &p4_pt_cut,
     const float &p1_eta_cut, const float &p2_eta_cut, const float &p3_eta_cut, const float &p4_eta_cut,
-    const int &p1_trigger_particle_id_cut, const int &p2_trigger_particle_id_cut, const int &p3_trigger_particle_id_cut, const int &p4_trigger_particle_id_cut,
+    const unsigned short &p1_trigger_particle_id_cut, const unsigned short &p2_trigger_particle_id_cut, const unsigned short &p3_trigger_particle_id_cut, const unsigned short &p4_trigger_particle_id_cut,
     const int &p1_triggerbit_cut, const int &p2_triggerbit_cut, const int &p3_triggerbit_cut, const int &p4_triggerbit_cut, const float &DeltaR_threshold) {
 
     auto triggermatch = [DeltaR_threshold, p1_pt_cut, p2_pt_cut, p3_pt_cut, p4_pt_cut, p1_eta_cut,
@@ -705,7 +705,7 @@ ROOT::RDF::RNode GenerateQuadTriggerORFlag(
                             const ROOT::Math::PtEtaPhiMVector &particle3_p4,
                             const ROOT::Math::PtEtaPhiMVector &particle4_p4,
                             ROOT::RVec<int> triggerobject_bits,
-                            ROOT::RVec<int> triggerobject_ids,
+                            ROOT::RVec<unsigned short> triggerobject_ids,
                             ROOT::RVec<float> triggerobject_pts,
                             ROOT::RVec<float> triggerobject_etas,
                             ROOT::RVec<float> triggerobject_phis) {
@@ -849,8 +849,8 @@ ROOT::RDF::RNode MatchDoubleTriggerObject(
     const std::string &triggerobject_pt, const std::string &triggerobject_eta,
     const std::string &triggerobject_phi, const float &p1_pt_cut,
     const float &p2_pt_cut, const float &p1_eta_cut, const float &p2_eta_cut,
-    const int &p1_trigger_particle_id_cut,
-    const int &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
+    const unsigned short &p1_trigger_particle_id_cut,
+    const unsigned short &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
     const int &p2_triggerbit_cut, const float &DeltaR_threshold) {
 
     auto triggermatch =
@@ -860,7 +860,7 @@ ROOT::RDF::RNode MatchDoubleTriggerObject(
          p2_triggerbit_cut](const ROOT::Math::PtEtaPhiMVector &particle1_p4,
                             const ROOT::Math::PtEtaPhiMVector &particle2_p4,
                             ROOT::RVec<int> triggerobject_bits,
-                            ROOT::RVec<int> triggerobject_ids,
+                            ROOT::RVec<unsigned short> triggerobject_ids,
                             ROOT::RVec<float> triggerobject_pts,
                             ROOT::RVec<float> triggerobject_etas,
                             ROOT::RVec<float> triggerobject_phis) {
@@ -929,14 +929,14 @@ ROOT::RDF::RNode MatchSingleTriggerObject(
     const std::string &triggerobject_id, const std::string &triggerobject_pt,
     const std::string &triggerobject_eta, const std::string &triggerobject_phi,
     const float &pt_cut, const float &eta_cut,
-    const int &trigger_particle_id_cut, const int &triggerbit_cut,
+    const unsigned short &trigger_particle_id_cut, const int &triggerbit_cut,
     const float &DeltaR_threshold) {
 
     auto triggermatch = [DeltaR_threshold, pt_cut, eta_cut,
                          trigger_particle_id_cut, triggerbit_cut](
                             const ROOT::Math::PtEtaPhiMVector &particle_p4,
                             ROOT::RVec<int> triggerobject_bits,
-                            ROOT::RVec<int> triggerobject_ids,
+                            ROOT::RVec<unsigned short> triggerobject_ids,
                             ROOT::RVec<float> triggerobject_pts,
                             ROOT::RVec<float> triggerobject_etas,
                             ROOT::RVec<float> triggerobject_phis) {
@@ -997,9 +997,9 @@ bool matchParticle(const ROOT::Math::PtEtaPhiMVector &particle,
                    ROOT::RVec<float> &triggerobject_etas,
                    ROOT::RVec<float> &triggerobject_phis,
                    ROOT::RVec<int> &triggerobject_bits,
-                   ROOT::RVec<int> &triggerobject_ids, const float &matchDeltaR,
+                   ROOT::RVec<unsigned short> &triggerobject_ids, const float &matchDeltaR,
                    const float &pt_cut, const float &eta_cut,
-                   const int &trigger_particle_id_cut,
+                   const unsigned short &trigger_particle_id_cut,
                    const int &triggerbit_cut,
                    const float &trigger_particle_pt_cut) {
     Logger::get("CheckTriggerMatch")->debug("Checking Triggerobjects");
@@ -1100,7 +1100,7 @@ ROOT::RDF::RNode GenerateSingleTriggerFlag(
     const std::string &triggerobject_id, const std::string &triggerobject_pt,
     const std::string &triggerobject_eta, const std::string &triggerobject_phi,
     const std::string &hltpath, const float &pt_cut, const float &eta_cut,
-    const int &trigger_particle_id_cut, const int &triggerbit_cut,
+    const unsigned short &trigger_particle_id_cut, const int &triggerbit_cut,
     const float &DeltaR_threshold, const float &trigger_particle_pt_cut) {
 
     auto triggermatch =
@@ -1108,7 +1108,7 @@ ROOT::RDF::RNode GenerateSingleTriggerFlag(
          triggerbit_cut, trigger_particle_pt_cut](
             bool hltpath, const ROOT::Math::PtEtaPhiMVector &particle_p4,
             ROOT::RVec<int> triggerobject_bits,
-            ROOT::RVec<int> triggerobject_ids,
+            ROOT::RVec<unsigned short> triggerobject_ids,
             ROOT::RVec<float> triggerobject_pts,
             ROOT::RVec<float> triggerobject_etas,
             ROOT::RVec<float> triggerobject_phis) {
@@ -1232,8 +1232,8 @@ ROOT::RDF::RNode GenerateDoubleTriggerFlag(
     const std::string &triggerobject_pt, const std::string &triggerobject_eta,
     const std::string &triggerobject_phi, const std::string &hltpath,
     const float &p1_pt_cut, const float &p2_pt_cut, const float &p1_eta_cut,
-    const float &p2_eta_cut, const int &p1_trigger_particle_id_cut,
-    const int &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
+    const float &p2_eta_cut, const unsigned short &p1_trigger_particle_id_cut,
+    const unsigned short &p2_trigger_particle_id_cut, const int &p1_triggerbit_cut,
     const int &p2_triggerbit_cut, const float &DeltaR_threshold,
     const float &p1_trigger_particle_pt_cut,
     const float &p2_trigger_particle_pt_cut) {
@@ -1247,7 +1247,7 @@ ROOT::RDF::RNode GenerateDoubleTriggerFlag(
                             const ROOT::Math::PtEtaPhiMVector &particle1_p4,
                             const ROOT::Math::PtEtaPhiMVector &particle2_p4,
                             ROOT::RVec<int> triggerobject_bits,
-                            ROOT::RVec<int> triggerobject_ids,
+                            ROOT::RVec<unsigned short> triggerobject_ids,
                             ROOT::RVec<float> triggerobject_pts,
                             ROOT::RVec<float> triggerobject_etas,
                             ROOT::RVec<float> triggerobject_phis) {
